requirements.txt
flask>=2.0.0
opencv-python
mediapipe
numpy
pillow
scikit-learn 